// src/interfaces/mongoose.gen.ts
export type CatalogCourseQueries = {
}

export type CatalogCourseMethods = {
}
// ...
export type CatalogCourseDocument = mongoose.Document<mongoose.Types.ObjectId, CatalogCourseQueries> & CatalogCourseMethods & {
active?: boolean;
aka?: string;
antireq?: any;
antireq_text?: string;
career?: string;
cid?: string;
code?: string;
components: mongoose.Types.Array<string>;
coreq?: any;
coreq_text?: string;
course_group_id?: string;
course_number?: string;
coursedog_id?: string;
units?: number;
departments: mongoose.Types.Array<string>;
description?: string;
faculty_code?: string;
faculty_name?: string;
grade_mode_code?: string;
grade_mode_name?: string;
long_name?: string;
multi_term?: boolean;
name?: string;
nogpa?: boolean;
notes?: string;
prereq?: any;
prereq_text?: string;
repeatable?: boolean;
start_term?: mongoose.Types.Map<any>;
subject_code?: string;
topics: mongoose.Types.Array<mongoose.Types.Map<any>>;
version?: number;
_id: mongoose.Types.ObjectId;
createdAt?: Date;
updatedAt?: Date;
}